Document a postback will contain the most recent data
Rephrased to avoid the suggestion that we will sent a postback with only the data appended for a certain activity while we will actually sent a full json with all the data already known to us. For example adding an activity can (and will) also include data for activities which were added immediatly after the current activity.