Document clarification activity 203
The signer activity 'signed' 203 only indicates that the signer finished
out signing flow with the intent of signing the document(s). After this
we still have some processing to do before the signed document is available,
this is when the transaction reaches status 30.

By documenting this more clearly we can hopefully avoid the scenario in which
users try to download the document based on the 203 activities of their signers
instead of the overall transaction signed (30) status.